Moves config functions to seperate file
